refactor(storybook): update from @storybook/react-webpack5 to @storybook/react-vite
