+++ packages/react/.storybook/preview.jsx
@@ -1,6 +1,6 @@
 import {PrimerBreakpoints} from '../src/utils/layout'
 import React, {useEffect} from 'react'
-import {ThemeProvider, BaseStyles, theme} from '../src'
+import {ThemeProvider, BaseStyles} from '../src'
 import {FeatureFlags} from '../src/FeatureFlags'
 import {DefaultFeatureFlags} from '../src/FeatureFlags/DefaultFeatureFlags'
 import {clsx} from 'clsx'
