+++ packages/react/package.json
@@ -143,10 +143,9 @@
     "@storybook/addon-interactions": "^8.3.0",
     "@storybook/addon-links": "^8.3.0",
     "@storybook/addon-storysource": "^8.3.0",
-    "@storybook/addon-webpack5-compiler-babel": "^3.0.3",
     "@storybook/preview-api": "^8.3.0",
     "@storybook/react": "^8.3.0",
-    "@storybook/react-webpack5": "^8.3.0",
+    "@storybook/react-vite": "^8.3.0",
     "@storybook/source-loader": "^8.3.0",
     "@storybook/theming": "^8.3.0",
     "@testing-library/dom": "^10.4.0",
@@ -167,7 +166,6 @@
     "ajv": "8.16.0",
     "axe-core": "4.9.1",
     "babel-core": "7.0.0-bridge.0",
-    "babel-loader": "^9.1.0",
     "babel-plugin-add-react-displayname": "0.0.5",
     "babel-plugin-dev-expression": "0.2.3",
     "babel-plugin-macros": "3.1.0",
@@ -179,7 +177,6 @@
     "concurrently": "8.0.1",
     "copyfiles": "2.4.1",
     "cross-env": "7.0.3",
-    "css-loader": "^7.1.2",
     "fast-glob": "3.3.2",
     "filesize": "10.1.2",
     "front-matter": "4.0.2",
@@ -199,7 +196,6 @@
     "micromark-extension-frontmatter": "2.0.0",
     "micromark-extension-mdxjs": "3.0.0",
     "mini-css-extract-plugin": "^2.9.1",
-    "postcss-loader": "^8.1.1",
     "postcss-preset-primer": "^0.0.0",
     "react": "18.3.1",
     "react-dnd": "14.0.4",
@@ -214,8 +210,6 @@
     "rollup-plugin-visualizer": "5.9.2",
     "semver": "7.6.2",
     "storybook": "^8.3.0",
-    "storybook-addon-turbo-build": "2.0.1",
-    "style-loader": "4.0.0",
     "styled-components": "5.3.11",
     "terser": "5.31.0",
     "ts-toolbelt": "9.6.0",
@@ -225,7 +219,6 @@
     "unist-util-find": "3.0.0",
     "unist-util-find-before": "4.0.0",
     "unist-util-flat-filter": "2.0.0",
-    "webpack": "5.94.0",
     "yaml": "2.3.4"
   },
   "peerDependencies": {
